@@ -11,15 +11,14 @@
     ],
     "minimum-stability": "stable",
     "require": {
-        "php": ">=7.1",
+        "php": ">=7.4",
         "ext-json": "*",
-        "guzzlehttp/guzzle": "~6.0",
         "psr/http-client": "^1.0",
-        "psr/http-factory": "^1.0",
-        "ricardofiorani/guzzle-psr18-adapter": "^1.0"
+        "psr/http-factory": "^1.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "~7.0"
+        "phpunit/phpunit": "~9.0",
+        "guzzlehttp/guzzle": "^7.3"
     },
     "autoload": {
         "psr-4": {
